+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## v0.1.2
+* Refactoring
+* Fix roles
+* Fix tests
+* Update rust version in docker build
+
 ## v0.1.1
 * Update dependencies
 * Minor refactoring